requirements.txt
flask
gunicorn
tensorflow
numpy
pillow
firebase-admin
google-cloud-storage
google-cloud-secret-manager